Refactor property names and types for consistency
Renamed properties for clarity: `EmployeeBadge` to `BadgeNumber` and `EmployeeSsn` to `Ssn`. Updated types: `BadgeNumber` to `int` and `Ssn` to `string` where applicable. Reflected changes across C# classes, TypeScript interfaces, and CSV mappings. Removed SSN value formatters in grid columns. Simplified test cases by removing unnecessary type casting. These changes enhance code readability and maintainability.

+++ src/services/tests/Demoulas.ProfitSharing.UnitTests/TotalServiceTests.cs
@@ -93,7 +93,7 @@ public class TotalServiceTests : ApiTestBase<Program>
             var ppTest = await ctx.PayProfits.FirstAsync(CancellationToken.None);
             ppTest.ProfitYear = 2100;
             ppTest.Etva = 4321;
-            var etvaResult = await _totalService.GetTotalEtva(ctx, (short)2100).FirstAsync();
+            var etvaResult = await _totalService.GetTotalEtva(ctx, 2100).FirstAsync();
             etvaResult.Total.Should().Be(4321);
         });
     }
